Implemented MOAITextRenderer::ProcessOptimalSize().  Changed type of MOAITextRenderer's static contstant to int. Made MOAIFreeTypeFont methods EstimatedMaxFontSize() and NumberOfLinesToDisplayText() public.  Added and implemented MOAIFreeTypeFont methods IsFreeTypeInitialized() and SetCharacterSize().
Conflicts:
	src/moaicore/MOAIFreeTypeFont.h

+++ src/moaicore/MOAITextRenderer.h
@@ -57,7 +57,7 @@ private:
 public:
 	DECL_LUA_FACTORY ( MOAITextRenderer )
 	
-	static const float PROCESSING_IN_PROGRESS = -1.0f;
+	static const int PROCESSING_IN_PROGRESS = -1;
 	
 	MOAITextRenderer ();
 	~MOAITextRenderer ();
